refactor: move handling views and hooks to own dirs
