+++ frontend/benefit/handler/src/components/applicationReview/handlingView/ApplicationStepper.sc.ts
