+++ frontend/benefit/handler/src/hooks/applicationHandling/useDecisionProposalTemplateQuery.ts
