+++ frontend/benefit/handler/src/hooks/applicationHandling/useHandlingStepper.ts
